Add whitelisted telemetry target for tracing.
