+++ packages/next/telemetry/trace/report/to-telemetry.ts
@@ -1,10 +1,27 @@
-const reportToTelemetry = (
-  _spanName: string,
-  _duration: number,
-  _timestamp: number,
-  _id: string | null,
-  _parentId?: string,
-  _attrs?: Object
-) => {}
+import { traceGlobals } from '../shared'
+
+const TRACE_EVENT_WHITELIST = new Map(
+  Object.entries({
+    'webpack-invalidated': 'WEBPACK_INVALIDATED',
+  })
+)
+
+const reportToTelemetry = (spanName: string, duration: number) => {
+  const eventName = TRACE_EVENT_WHITELIST.get(spanName)
+  if (!eventName) {
+    return
+  }
+  const telemetry = traceGlobals.get('telemetry')
+  if (!telemetry) {
+    return
+  }
+
+  telemetry.record({
+    eventName,
+    payload: {
+      durationInMicroseconds: duration,
+    },
+  })
+}
 
 export default reportToTelemetry
