--- packages/next/telemetry/trace/types.ts
@@ -1,7 +1,0 @@
-export enum TARGET {
-  CONSOLE = 'CONSOLE',
-  ZIPKIN = 'ZIPKIN',
-  TELEMETRY = 'TELEMETRY',
-}
-
-export type SpanId = string
